pyproject.toml
[project]
name = "chat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.3.2",
    "python-dotenv>=1.0.1",
    "google-generativeai>=0.8.5",
    "fastapi>=0.115.12",
    "fastapi-mcp>=0.3.4",
    "httpx>=0.27.0",
    "uvicorn>=0.34.2",
    "pillow>=11.2.0",
    "google-genai>=1.17.0",
]
